Projectiles can hit players and tiles. Damage dealt depends on base damage and distance from impact. Also major cleanup and more parameters are passed to projectile updates so there's not need for weird spell functions for teleport and future things that interact with players. Long message__wow

+++ Game/Globals.h
@@ -17,7 +17,7 @@ extern const int TILES_Y;
 extern const float PI;
 extern const float PI_OVER_180;
 
-std::string ToString(int);
+std::string ToString(float);
 std::string ToString(sf::Vector2i);
 void DrawDebugString();
 void CreateDebugString();
